Adding MiniMax in Ver03
The Minimax Algorithm:
Recall the minimax algorithm originally presented at the beginning of this lesson (shown below). We will start by coding helper functions: a terminal state test, the min-value function, and the max-value function. Notice that the min-value() and max-value() functions only need to return a single number -- a "utility value" -- which is the score of the current branch; the function for selecting a move based on those scores will come in a later quiz.